\name{add_keybinding}
\alias{add_keybinding}
\title{Add keybinding to document for this action. Key is value for
Ext.EventObject:
http://docs.sencha.com/ext-js/4-1/#!/api/Ext.EventObject. Use
Ctrl-X, Alt-X of Shift-X indicate keys}
\usage{
  add_keybinding(obj, key.accel)
}
\description{
  Add keybinding to document for this action. Key is value
  for Ext.EventObject:
  http://docs.sencha.com/ext-js/4-1/#!/api/Ext.EventObject.
  Use Ctrl-X, Alt-X of Shift-X indicate keys
}

