\name{call_ext}
\alias{call_ext}
\title{call an ext method}
\usage{
  call_ext(obj, meth, params = "")
}
\arguments{
  \item{obj}{object}

  \item{meth}{method name}

  \item{params}{optional parameters. You must format as
  string}
}
\value{
  pushes value to queue, nothing for you to do
}
\description{
  call an ext method
}

