\name{enabled<-}
\alias{enabled<-}
\alias{enabled<-.default}
\title{Set if widget is enabled}
\usage{
  enabled(obj) <- value

  \method{enabled}{default}(obj) <- value
}
\arguments{
  \item{x}{widget}

  \item{value}{logical}
}
\description{
  Set if widget is enabled

  assignment method for enabled
}

