\name{gcombobox}
\alias{gcombobox}
\title{comobobox}
\usage{
  gcombobox(items, selected = 1, editable = FALSE,
    coerce.with = NULL, handler = NULL, action = NULL,
    container = NULL, ..., width = NULL, height = NULL,
    ext.args = NULL, autocomplete = FALSE,
    initial.msg = "", tpl = NULL)
}
\description{
  comobobox
}

