\name{gexpandgroup}
\alias{gexpandgroup}
\title{expandable box container}
\usage{
  gexpandgroup(text = "", horizontal = TRUE, spacing = 2,
    use.scrollwindow = FALSE, container = NULL,
    handler = NULL, action = NULL, ..., width = NULL,
    height = NULL, ext.args = NULL)
}
\description{
  Use \code{visible<-} to toggle collapsed or expanded
  programatically. Opens in a visible mode.  Use
  \code{addHandlerChanged} to add a callback for when the
  widget expands or collapses
}

