\name{gframe}
\alias{gframe}
\title{Framed box container}
\usage{
  gframe(text = "", horizontal = TRUE, spacing = 2,
    body.padding = 2, use.scrollwindow = FALSE,
    container = NULL, ..., width = NULL, height = NULL,
    ext.args = list())
}
\description{
  Use \code{svalue<-} to adjust title
}

