\name{gstatusbar}
\alias{gstatusbar}
\title{Status bar for gwindow instances}
\usage{
  gstatusbar(text = "", container = NULL, ...,
    ext.args = NULL)
}
\arguments{
  \item{text}{text for label}
}
\description{
  Status for main window. Use \code{gwindow} instance for
  parent container. The \code{svalue<-} method can be used
  to change the value.
}

