\name{gtable}
\alias{gtable}
\title{gtable}
\usage{
  gtable(items, multiple = FALSE, chosencol = 1,
    icon.col = NULL, tooltip.col = NULL, handler = NULL,
    action = NULL, container, ..., width = NULL,
    height = NULL, ext.args = list(), store.args = list(),
    selection = if (multiple) "multiple" else "single",
    paging = NULL, col.widths = 1)
}
\arguments{
  \item{selection}{one of 'single', 'multiple', or
  'checkbox'. Defaults to choice of multiple. The
  'checkbox' options gives intuitive checkboxes for
  selection.}
}
\description{
  gtable
}

