\name{addHandlerChanged}
\alias{addHandlerBlur}
\alias{addHandlerChange}
\alias{addHandlerChanged}
\alias{addHandlerClicked}
\alias{addHandlerDoubleclick}
\alias{addHandlerEnter}
\alias{addHandlerMouseDown}
\alias{addHandlerMouseUp}
\alias{addHandlerSelect}
\title{Name given to most common action (widget changed state, ...)}
\usage{
  addHandlerChanged(obj, handler, action = NULL, ...)

  addHandlerClicked(obj, handler, action = NULL, ...)

  addHandlerDoubleclick(obj, handler, action = NULL, ...)

  addHandlerEnter(obj, handler, action = NULL, ...)

  addHandlerBlur(obj, handler, action = NULL, ...)

  addHandlerSelect(obj, handler, action = NULL, ...)

  addHandlerChange(obj, handler, action = NULL, ...)

  addHandlerMouseDown(obj, handler, action = NULL, ...)

  addHandlerMouseUp(obj, handler, action = NULL, ...)
}
\description{
  Name given to most common action (widget changed state,
  ...)

  Handler for a click event

  Handler for a double click event

  Handler for the enter key event?

  handler for loss of keyboard focus event

  handler for select event

  handler for change of state event

  handler for mouse down event

  handler for mouse up event
}

