\name{set_button_icon}
\alias{set_button_icon}
\title{Set the icon for this button}
\usage{
  set_button_icon(obj, value)
}
\arguments{
  \item{obj}{button object}

  \item{value}{stock icon name}
}
\description{
  Called by svalue, but can also be called directly to
  override default which matches by button label.
}

