\name{set_panel_collapse}
\alias{set_panel_collapse}
\title{collapse or expand (collapse=FALSE) collapsible panel}
\usage{
  set_panel_collapse(obj,
    where = c("center", "north", "south", "east", "west"),
    collapse = TRUE)
}
\description{
  collapse or expand (collapse=FALSE) collapsible panel
}

