\name{set_panel_size}
\alias{set_panel_size}
\title{set theh panel size}
\usage{
  set_panel_size(obj,
    where = c("center", "north", "south", "east", "west"),
    dimension)
}
\arguments{
  \item{where}{which panel (center, north, ...),}

  \item{dimension}{width or height as appropriate}
}
\description{
  set theh panel size
}

