\name{size<-}
\alias{size<-}
\alias{size<-.default}
\title{Set size property, if implemented}
\usage{
  size(obj) <- value

  \method{size}{default}(obj) <- value
}
\arguments{
  \item{x}{widget}

  \item{...}{passed on}

  \item{value}{size specification, for most widgets a pair
  c(width, height), but can have exceptions}
}
\description{
  Set size property, if implemented

  assignment method for size
}

