\name{visible}
\alias{visible}
\alias{visible.default}
\title{Generically returns if widget is in visible state}
\usage{
  visible(obj)

  \method{visible}{default}(obj)
}
\arguments{
  \item{x}{object Calls objects \code{set_visible} method}
}
\description{
  Generically returns if widget is in visible state

  getters method for visible
}

