\name{visible<-}
\alias{visible<-}
\title{Primarily used to set if widget is shown, but also has other meanings}
\usage{
  visible(obj) <- value
}
\arguments{
  \item{x}{object Calls objects \code{set_visible} method}

  \item{value}{logical}
}
\description{
  Primarily used to set if widget is shown, but also has
  other meanings
}

