\name{ggroup}
\alias{ggroup}
\title{The basic box container}
\usage{
  ggroup(horizontal = TRUE, spacing = 2, body.padding = 2,
    use.scrollwindow = FALSE, container = NULL, ...,
    width = NULL, height = NULL, ext.args = list(),
    align = "stretch", text = NULL, collapsible = FALSE)
}
\description{
  Packing is from left to right (horizontal) or top to
  bottom (gvbox)
}

