\name{gtable}
\alias{gtable}
\title{gtable}
\usage{
  gtable(items, multiple = FALSE, chosencol = 1,
    icon.col = NULL, tooltip.col = NULL, handler = NULL,
    action = NULL, container, ..., width = NULL,
    height = NULL, ext.args = list(), store.args = list(),
    selection = TRUE, paging = NULL, col.widths = 1)
}
\description{
  gtable
}

